Smart-Finance-Manager/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-extensions==4.1
gunicorn==22.0.0
whitenoise==6.6.0
idna==3.11
packaging==25.0
pillow==12.0.0
pytesseract==0.3.13
reportlab==4.4.4
requests==2.32.5
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
